refactor: cleaned the useOnboardingData hook
